fix: unit tests for usePipelinerunsV2
